Make @theia/electron a development dependency
Fixes #48

ATM @theia/electron is a runtime dependency, that depends on the
"electron" package. As such both are preserved with the packaged
blueprint application.

Simple solution: make @theia/electron a dev-dependency, This way
it will be stripped when we package blueprint, somewhat reducing
our various package sizes. On Linux the expected size reduction
is 16MB.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>