add Gitpod config
ATM it's difficult to produce installer for test purposes in a
reproducable way, e.g. to test PRs. Though not mandatory for
anyone to use, Gitpod is use by several committers and permits
programmatically generating a workspace that contains the
Linux packages, e.g. Appimage, that can be started in the Gitpos
workspace (using VNC since it's an Electron app).

The Gitpod config was taken from eclipse-theia/theia and slightly
adapted for this repo. I have added the generation of preview and
production packages for Linux. This provides an easy means of testing
and looking under the hood, e.g. for troubleshooting, in an
environment that should offer reproducibility.

For enhanced benefit we can consider having webmaster enable the
Gitpod integration for this repo. Then we'd have automatic workspaces
pre-bbuils for PRs.

To open this PR in Gitpod and look-around, either use their browser
add-on or prepend the following URL to the GitHub URL of this PR:
"https://gitpod.io/#" .

Once merged, this trick will work with any future branch/PR/commit
in this repo, since they'll have a Gitpod config.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>