replace vscode built-ins by extension pack
We now have an extension pack, that lists all vscode-builtin
extensions.

As a first step, let's aim to have more features than
not, and use that pack. For the most part, extensions
listed in a pack should be installed, at the latest
version that is compatible with the currently used Theia
framework version, wr to the vscode extensions API. We
need not manually update the version of the built-in
extensions, any more.

A few of the vscode built-in extensions do not work well,
with the version pulled by the pack. These we need to pin to a
specific version, that's known to work, using the old way.
e.g.:
`vscode.git` requires v1.52.1 instead of v1.50.0

Closes #26

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>