Delete content of Jenkins job cache folders
Directories  /.cache/ and /.electron-gyp/ are defined
in the jenkins config in section volumeMounts. As per
documentation they are persistent for the life of the
pod. I think once in a while a problem happened with
electron-gyp, that made the job un-buildable from that
moment-on.

This PR was in that state, where nothing would clear
the gyp errors, until I tried cleaning the directory,
which apparently did the trick.

To avoid the same issue in the future, we may as well
keep that line.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>