Fix Jenkins CI
Jenkins has been failing recently, following an attempt
at improving it, by clearing the caches. I noticed some
signs that the /.cache folder is used for more than yarn,
and so should probably not be wiped althogether.

This commit rolls-back the deletion of /.cache and instead
relies on existing "yarn cache clean" to do that job.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>