[ci] Proceed with signing only during Jenkins CI on Mac
ATM we have a post-packaging step that proceeds to sign
binaries from the Mac package. This can only work during
Jenkins CI, since it uses Eclipse Foundation infrastructure,
which is only configured for that context.

Since no distinction is made whether the context is CI or
something else, the same happens whenever someone on a Mac
attempts to follow the README instructions to try-out
building and locally packaging Blueprint. This can lead to
confusion and a generally bad experience.

This PR sets-up an environment variable in the Jenkins
config, and makes signing conditional to the presence of
this variable.

Fixes #138

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>