Remove "Skip Mac CI signing" printout if not on Mac
This is a follow-up on the recent commit that made signing
on Mac happen only when running Jenkins CI. The printout
above was added for all operating systems, but it really
only makes sense to have it when one is running on a Mac.

This commit removes the printout on other platforms.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>