Streamline project's CI
Fixes #148

ATM Jenkins runs every time a PR branch is modified. Final
Mac package signing occurs only when merging to master, but
signing of various binaries happens all the time. It's not
obvious how to avoid it - it's a post-packaging step - while
having the Jenkins run be meaningful.

Also, we have recently added a GitHub workflow that gives us
rapid feedback about updated PRs on our 3 target platforms.

Why not use the GH workflow to build and test packages for our
various operating systems, and give Jenkins executor a break
by running that CI only when needed?

I propose we:

- enhance the GH workflow to generate the packages, without
setting-up the environment variable that enables signing
binaries on Mac. We can then run our modest test suite on
these packages and have an idea if we broke something.
- only run Jenkins CI when merging on master or when we
have reasonable suspicion(1) that the PR changes the Jenkins
configuration, in which case we probably want Jenkins to
run as it does today on PRs, detecting update issues,
without having to merge to master first.

This will save a lot of CI time and insure Jenkins executor
are available when we really need them: when we merge to master
and want new packages generated, signed and uploaded.

(1) when the word "jenkins" appears in the PR title
or PR branch name.

Note: a slight change in the Jenkins multi-branch pipeline
configuration (from server web UI) was necessary for this to
work. So far, We had one pipeline branch configured, triggered
by either a merge to master or a PR branch change. In order
for the git branch to be programmatically visible from the
`Jenkinsfile`, it was necessary to split these into 2 separate
pipelines branch configuration, each one "watching" for a single
thing (master branch, PR-related branch).
Ref: https://issues.jenkins.io/browse/JENKINS-7554

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>