[CI] follow-up: fix CI when merging
This is a follow-up to #143. After merging, I noticed that
Jenkins CI run was unexpectedly streamlined:
https://ci.eclipse.org/theia/job/Theia2/job/master/5/console

Looking at that log, it seems the environment available from
Jenkinsfile is a bit different, when merging vs when it's a PR.
Variable "env.CHANGE_BRANCH" is not set when merging,
so the "when" conditions did not work as expected.

Looking at a recent log for master, it seems "env.GIT_BRANCH"
is set and has correct value, so let's try using that instead.
log: https://ci.eclipse.org/theia/job/Theia2/job/master/4/consoleFull

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>