[Updater] Allow to inspect logs on error
Add an action to the error notification that allows the user to see the
log file with the full error message.

The path to the error log is retrieved from the logger that was used to
record the error. The path must be given as part of the error object for
the action button to appear.

Signed-off-by: Alexandra Buzila <abuzila@eclipsesource.com>