Set fixed version for eslint-plugin-deprecation
The recent update of eslint-plugin-deprecation pulls in a newer version
of @typescript-eslint/experimental-utils.
This version is not compatible with the node versions available on the
shared windows and mac build nodes. (For Linux we could update the
docker image):

error @typescript-eslint/experimental-utils@5.8.0: The engine "node" is
incompatible with this module. Expected version "^12.22.0 || ^14.17.0 ||
>=16.0.0". Got "12.20.0"

For now set eslint-plugin-deprecation to the version known to work.