pin colors dependency
The latest version of this package has apparently been
sabotaged by the author. Upstream theia is replacing it,
but since it's also an indirect dependency, we have to
pin it, so that the latest version is not used. Else we'll
have problems, e.g. each time we run `yarn upgrade`.

see:
https://community.theia-ide.org/t/colors-js-module-is-sabotaged-by-own-developer/2173
https://github.com/eclipse-theia/theia/discussions/10610

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>