Clarify update dialogs
Previously, the update dialogs gave the impression that a downloaded
update wouldn't get installed if answering 'No' in the second dialog.
Since this isn't the case, make it clear in the message that the update
will get installed on application exit and that the question is whether
to restart now or not.

Fixes #212

Contributed by STMicroelectronics

Signed-off-by: Erik LUNDIN <erik.lundin@st.com>