[Gitpod] update config to use node 14
I find Gitpod is useful for quickly testing some PRs. ATM the node version
used is not compatible with Theia and so the build fails. This commit updates
the config to use node 14.

Also start the electron version of Blueprint when the build is done. It's
accessible from the Ports view (bottom bar) by selecting "open browser"
in line with "port 6080".

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>