A few tweaks to the new "cleanup" Jenkins file
Looking at the run of the first run, I noticed that seemingly no
`yarn--<random number>` or `theia-plugin-download<random identifier>`
files were found or cleaned. It was confirmed in the Gitlab ticket that
these two types of files are still present on the server.

In this commit:
- adapt the pattern we use for those two file types
- handle the case of the "theia-plugin-download**" files differently,
looking-for and deleting individual files that match the pattern rathern
than directories.
- various little improvements, guided by linter feedback

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>