[built-ins] Update bundled extensions
Since Theia recently has caught-up to vscode 1.77.0,  we are now able
to use the latest version of pretty much every extension, in Blueprint.

Doing so means a massive update, version-wise, of the vscode builtins and
Java extensions that we bundle. So, it might bring some little issues.
But it will also bring big benefits.

I have removed all pinned version of built-ins, because I beleive we can
successfully run the latest of everything.

I have also replaced individual Java extensions with a pack, which
dependends on the few most popular Java extensions. One of them,
I beleive, is proprietary: `Visual Studio IntelliCode`. I've added
it to the "exclude list". In anycase, this extension is not available
on the public open-vsx registry and so can't be fetched by mistake,
e.g. if the exclusion for it were removed by accident.

Done like this, it should be more future-proof: we will pick the latest
compatible version of each bundled extension, each time we build. If
we encounter punctual problems in the future with later versions of the
extensions we bundle, we can use strategic temporary pinning, to
maintain functionality while fixing the issue.

[build] Make the Theia plugin download script go at it sequentially

I've had the plugins download fail on occasion locally, over
"redhat.java". I suspect a race condition, when using the default
mode where multiple plugins are downloaded at the same time.

[Theia] add `@theia/secondary-window`

I've added this extension to the browser version of Blueprint. We
can add it to the Electron version of Blueprint once it's available
in a release.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>