[js/ts][debug] temporarily add js-debug* extensions
Following the example of the upstream Theia example application:

"Due to an issue when publishing the builtin extension-pack for `1.79.0`
we need to manually add the `js-debug` and `js-debug-companion` plugins
for JavaScript debugging support."

see: https://github.com/eclipse-theia/theia/pull/12835

This commit can be reverted when Blueprint uses a newer release of
vscode-built-in extension pack, that includes the js-debug*
extensions. No such new release exists as of this writing.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>