Updated docker base image to node:18-bullseye in browser.Dockerfile
Resolves https://github.com/eclipse-theia/theia-blueprint/issues/345

`node:16-bullseye` seems outdated and resulted in the bug documented in the above issue.  It is not currently supported on the official list of node docker images:
https://hub.docker.com/_/node

The issue at https://github.com/eclipse-theia/theia-blueprint/issues/345 was resolved even by changing to `node:18-bullseye` only in the `build-stage`, but presumably it should be already upgraded also in the `production-stage`, so this commit/PR also updates the production stage based image.