Updated the "What is this?" section of the Theia IDE README
This commit's primary purpose is to suggest and/or clarify Theia IDE's purpose as more than merely a 'template' for custom applications based on the Theia platform, but also --- indeed, primarily --- as a 'minimum viable product' that can be used out-of-the-box.  (Which should be supported by the development team as such.)

Along the way this commit reorganizes and improves this entire paragraph/section.  Particularly:

This commit also added a missing hyperlink to the Theia.cloud demo.

This commit removed the first sentence ("The Eclipse IDE is a modern and open IDE for cloud and desktop"), as in my opinion it may only add further confusion to a new user and the focus here in the README of the Theia IDE should be a clear communication of the purpose of *this* product (not the Eclipse IDE), and of the distinction and relation between Theia IDE and the Theia platform.  (If it is truly important here to mention the Eclipse IDE I suggest that this sentence should be written more clearly, and also relocated e.g. to the end of the paragraph, rather than being the very first sentence.)

This commit relaxed the excessive '-based products based on' language.