Update README.md build steps with note to start clean after dependency changes/pulling to a new commit
Proposing a note in the readme for those doing manual builds to run `git clean -xfd` when updating to a newer commit of the repo, particularly when updating dependencies. 

The jenkinsfile already does this, see https://github.com/eclipse-theia/theia-blueprint/issues/349#issuecomment-2364956584 for more info. As a casual hobbyist developer who comes back to working on Theia intermittently this issue has come back to frustrate me for a while and seems to have come up in other conversations. There may be some better ways to handle this but the `git clean -xfd` approach has seemingly been more reliable.